moved core out of __init__
